+++ src/frontend/js/data/useCourseSearch/index.ts
@@ -1,7 +1,8 @@
 import { stringify } from 'query-string';
-import { UseQueryOptions, useQuery } from 'react-query';
+import { useQuery, UseQueryOptions } from 'react-query';
 
 import { APIListRequestParams } from 'types/api';
+import useLocalizedQueryKey from 'utils/react-query/useLocalizedQueryKey';
 import { fetchList, FetchListResponse } from '../getResourceList';
 
 export const useCourseSearch = (
@@ -10,16 +11,20 @@ export const useCourseSearch = (
     FetchListResponse,
     unknown,
     FetchListResponse,
-    (string | APIListRequestParams)[]
+    readonly (string | APIListRequestParams)[]
   > = {},
 ) => {
-  const queryKey = ['courses', stringify(searchParams)];
-  return useQuery<FetchListResponse, unknown, FetchListResponse, (string | APIListRequestParams)[]>(
-    queryKey,
-    async () => fetchList('courses', searchParams),
-    {
-      keepPreviousData: true,
-      ...queryOptions,
-    },
-  );
+  const queryKey = useLocalizedQueryKey(['courses', stringify(searchParams)]) as readonly (
+    | string
+    | APIListRequestParams
+  )[];
+  return useQuery<
+    FetchListResponse,
+    unknown,
+    FetchListResponse,
+    readonly (string | APIListRequestParams)[]
+  >(queryKey, async () => fetchList('courses', searchParams), {
+    keepPreviousData: true,
+    ...queryOptions,
+  });
 };