+++ CHANGELOG.md
@@ -61,6 +61,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
+- Make useCourseSearch hook locale sensitive
 - Stop using {% blockplugin %} template tags in <header> and replace them by
   simple {% if %} tags that do the same and don't inject frontend editing markup
 - Fix `get_placeholder_plugins` when placed in <header> by refactoring and